+++ cms/djangoapps/contentstore/config/waffle.py
@@ -6,7 +6,7 @@ waffle switches for the contentstore app.
 
 from edx_toggles.toggles import WaffleFlag, WaffleSwitch
 
-from openedx.core.djangoapps.waffle_utils.__future__ import FutureCourseWaffleFlag as CourseWaffleFlag
+from openedx.core.djangoapps.waffle_utils import CourseWaffleFlag
 
 # Namespace
 WAFFLE_NAMESPACE = 'studio'
