+++ lms/djangoapps/discussion/toggles.py
@@ -1,7 +1,7 @@
 """
 Discussions feature toggles
 """
-from openedx.core.djangoapps.waffle_utils.__future__ import FutureCourseWaffleFlag as CourseWaffleFlag
+from openedx.core.djangoapps.waffle_utils import CourseWaffleFlag
 
 WAFFLE_FLAG_NAMESPACE = "discussions"
